# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend cho ứng dụng SmartSchedule - Lịch hẹn thông minh",
        "title": "SmartSchedule API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@smartschedule.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:7070",
    "basePath": "/api/v1",
    "paths": {
...
        "/private/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về danh sách thông báo của người dùng hiện tại",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Lấy danh sách thông báo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Số trang",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số lượng mỗi trang",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/notifications/mark-all-read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Đánh dấu tất cả thông báo của người dùng là đã đọc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Đánh dấu tất cả đã đọc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/notifications/mark-read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Đánh dấu các thông báo cụ thể là đã đọc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Đánh dấu đã đọc",
                "parameters": [
                    {
                        "description": "Danh sách ID thông báo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MarkAsReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về số lượng thông báo chưa đọc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Đếm thông báo chưa đọc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.MarkAsReadRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                2006,
                2007,
                2008,
                2009,
                2010,
                3011,
                3012,
                3013,
                3014,
                3015,
                3016,
                3017,
                3018,
                3019,
                3020,
                3021,
                4022,
                4023,
                4024,
                4025,
                5026,
                5027,
                5028,
                5029,
                6030,
                6031,
                6032,
                6033,
                7034,
                7035
            ],
            "x-enum-varnames": [
                "ErrInvalidCredentials",
                "ErrTokenExpired",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrMissingAuthorizationHeader",
                "ErrInvalidTokenFormat",
                "ErrInvalidInput",
                "ErrInvalidEmail",
                "ErrInvalidPassword",
                "ErrInvalidFormat",
                "ErrInvalidRequestData",
                "ErrNotFound",
                "ErrAlreadyExists",
                "ErrResourceLocked",
                "ErrResourceExpired",
                "ErrCreateFailed",
                "ErrGetFailed",
                "ErrInvalidSlug",
                "ErrDuplicateSlug",
                "ErrInvalidParent",
                "ErrUpdateFailed",
                "ErrDeleteFailed",
                "ErrDatabase",
                "ErrDatabaseTimeout",
                "ErrUniqueViolation",
                "ErrForeignKey",
                "ErrBusinessRule",
                "ErrInvalidState",
                "ErrLimitExceeded",
                "ErrOperationFailed",
                "ErrInternalServer",
                "ErrConfiguration",
                "ErrThirdParty",
                "ErrNetwork",
                "ErrUploadToR2",
                "ErrSaveStorageDb"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}